+++ accountsignin.html
@@ -38,11 +38,11 @@
             </tr>
             <tr>
                 <td><button type="submit" id="submit">Submit</button></td>
+                <td><button onclick="deleteCookie()">Sign Out</button></td>
             </tr>
         </table>
         <p id="error"></p>
         <script src="/js/accountsigninjs.js"></script>
-        <script src="/js/accountjs.js"></script>
     </body>
     <!--Feet-->
 </html>
