--- js/accountjs.mjs
@@ -1,29 +1,0 @@
-export function setCookie(cname,cvalue,exdays) {
-  const d = new Date();
-  d.setTime(d.getTime() + (exdays*24*60*60*1000));
-  let expires = "expires=" + d.toGMTString();
-  document.cookie = cname + "=" + cvalue + ";" + expires + ";path=/";
-}
-export function getCookie(cname) {
-let name = cname + "=";
-let decodedCookie = decodeURIComponent(document.cookie);
-let ca = decodedCookie.split(';');
-for(let i = 0; i < ca.length; i++) {
-  let c = ca[i];
-  while (c.charAt(0) == ' ') {
-  c = c.substring(1);
-  }
-  if (c.indexOf(name) == 0) {
-  return c.substring(name.length, c.length);
-  }
-}
-return "";
-}
-export function checkCookie() {
-  let user = getCookie("email");
-  if (user != "" || user != null) {
-    alert("Welcome again " + user);
-  } else{
-    window.alert("test")
-  }
-}
